Rename filenames, dir names 'IDE' -> 'workspace'
1. Files that reference ide in their names have been renamed to workspace
2. Directories that reference ide in their names have been renamed to workspace
3. Imports have been changed appropriately
