+++ src/containers/workspace/ReplControlContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
 import { bindActionCreators, Dispatch } from 'redux'
 
 import { chapterSelect, clearReplOutput, evalEditor, evalRepl } from '../../actions/playground'
-import ReplControl, { IReplControlProps } from '../../components/IDE/ReplControl'
+import ReplControl, { IReplControlProps } from '../../components/workspace/ReplControl'
 import { IState } from '../../reducers/states'
 
 type DispatchProps = Pick<IReplControlProps, 'handleEvalRepl'> &
