+++ src/styles/_workspace.scss
