+++ src/styles/index.scss
@@ -5,6 +5,6 @@
 @import "global";
 @import "application";
 @import "blueprint";
-@import "ide";
+@import "workspace";
 @import "navigationBar";
 @import "playground";
